requirements.txt
python-telegram-bot==20.7
coinbase==2.1.0
aiosqlite==0.19.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
requests==2.31.0
